backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
PyJWT
email-validator
python-multipart